project-1/requirements.txt
attrs==21.2.0
bracex==2.1.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
click-option-group==0.5.3
colorama==0.4.4
coverage==6.0
dataclasses==0.6
Deprecated==1.2.13
Flask==2.0.2
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==21.0
peewee==3.14.4
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.19.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
semgrep==0.67.0
six==1.16.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcmatch==8.2
Werkzeug==2.0.2
wrapt==1.12.1
zipp==3.6.0
